blockchain, utreexobackends: move cached leaves map to utreexobackends
Moving cached leaves map to a separate package ensures that private
fields do not get accessed. This achieves better concurrency guarantees
as the exported functions of cached leaves map are concurrency safe but
if the fields get accessed directly, it bypasses the safety measures.
